Enemy Wave Spawning & Beginning of Enemy Path Implementation
Added Enemy and WaveSpawner classes to the project. The Enemy script is attached to an Enemy prefab and will keep track of its waypoints and have a method to move towards them. WaveSpawner is attached to a GameManager empty object in the scene and it spawns increasingly greater amounts of enemy waves at a certain spawn point with wait times in between wave spawning.

Current Caveats:
-When high enemy number waves are spawned enemies will overlap
-Current movement of enemies from waypoint to waypoint is not complete and in-progress...is open to changing for a different system as well.