Finished Enemy Movement & Path Following
Player 1 and Player 2  enemy types are implemented and moving in the pattern as discussed in meeting. To implement different types of enemies make sure to make a prefab that you pass into WaveSpawner.cs and you will have to write code to have it instantiated in the way you would like within WaveSpawner.cs. Currently the WeakEnemy type is used.